Facial_Expression_Recognition/requirements.txt
opencv-python
tensorflow
pillow
numpy